Simplify loading and storing integer functions.
For load32 and load64, we can simply delete the "optimized" case. For
little-endian targets, both GCC and Clang recognize the load of a native 32-bit
or 64-bit integer and turn it into the smallest possible move instruction.

Unfortunately, the same doesn't hold for stores, and we still need to detect the
native endianess. There's no standard way to do this, but using a constant union
seems to do the trick and is completely constant-folded away during compilation.

Remove the ALIGNED_ACCESS_REQUIRED configure check. This isn't a very good thing
to check through configure because it requires running an executable, which
doesn't work for cross compiles. Compilers generally know whether a target
supports unaligned accesses and if so, whether it's profitable to do in a
particular place. Even if the target doesn't support unaligned loads, the load &
store functions may be inlined somehwere will where higher alignment is
proven. My approach is to memcpy the integer through a union. Smart compilers
can see through the memcpy and turn it into a single move instruction if
possible.