Use "test x = y" instead of == in configure.ac
The ==  form is non-portable and not supported everywhere. Here is the explanation for this issue from pkgsrc:

The "test" command, as well as the "[" command, are not required to know                                                            
the "==" operator. Only a few implementations like bash and some
versions of ksh support it.

When you run "test foo == foo" on a platform that does not support the                                                              
"==" operator, the result will be "false" instead of "true". This can                                                               
lead to unexpected behavior.

There are two ways to fix this error message. If the file that contains                                                             
the "test ==" is needed for building the package, you should create a                                                               
patch for it, replacing the "==" operator with "=". If the file is not                                                              
needed, add its name to the CHECK_PORTABILITY_SKIP variable in the                                                                  
package Makefile.